Move `preprocess_forms` logic into its own file
